fix: style: Replace BootNetwork logo image. Modify gradient file. Rename SignupForm.
